requirements.txt
apache-airflow==2.9.0
openai
requests
